Fixes to Windows bugs and warnings

+++ src/tests/resource_manager_test.cpp
@@ -26,13 +26,13 @@ struct ResourceManagerFixture {
 int _test_loader(const char*, void* data, Resource* resource) {
     resource->i = 98;
     int* i = (int*)data;
-    (*i) += resource->i;
+    (*i) += (int)resource->i;
 
     return 0;
 }
 void _test_unloader(Resource* resource, void* data) {
     int* i = (int*)data;
-    (*i) -= resource->i;
+    (*i) -= (int)resource->i;
 }
 TEST_FIXTURE(ResourceManagerFixture, SetLoader) {
     int test_int = 0;
